+++ src/main/java/com/nisovin/shopkeepers/shopobjects/block/AbstractBlockShopObject.java
@@ -10,4 +10,18 @@ public abstract class AbstractBlockShopObject extends AbstractShopObject impleme
 	protected AbstractBlockShopObject(AbstractShopkeeper shopkeeper, ShopCreationData creationData) {
 		super(shopkeeper, creationData);
 	}
+
+	@Override
+	public abstract AbstractBlockShopObjectType<?> getType();
+
+	@Override
+	public boolean isActive() {
+		return (this.getBlock() != null);
+	}
+
+	@Override
+	public String getId() {
+		// Returns null if the block is null:
+		return this.getType().createObjectId(this.getBlock());
+	}
 }